+++ src/filters/Invert.ts
@@ -2,28 +2,10 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
+import { fragmentSource } from './invert.shaders';
 
 export const invertDefaultValues: Partial<TClassProperties<Invert>> = {
   alpha: false,
-  fragmentSource: `
-    precision highp float;
-    uniform sampler2D uTexture;
-    uniform int uInvert;
-    uniform int uAlpha;
-    varying vec2 vTexCoord;
-    void main() {
-      vec4 color = texture2D(uTexture, vTexCoord);
-      if (uInvert == 1) {
-        if (uAlpha == 1) {
-          gl_FragColor = vec4(1.0 - color.r,1.0 -color.g,1.0 -color.b,1.0 -color.a);
-        } else {
-          gl_FragColor = vec4(1.0 - color.r,1.0 -color.g,1.0 -color.b,color.a);
-        }
-      } else {
-        gl_FragColor = color;
-      }
-    }
-    `,
   invert: true,
   mainParameter: 'invert',
 };
@@ -67,6 +49,10 @@ export class Invert extends BaseFilter {
         data[i + 3] = 255 - data[i + 3];
       }
     }
+  }
+
+  protected getFragmentSource(): string {
+    return fragmentSource;
   }
 
   /**
