+++ src/filters/Pixelate.ts
@@ -2,29 +2,11 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
+import { fragmentSource } from './pixelate.shaders';
 
 export const pixelateDefaultValues: Partial<TClassProperties<Pixelate>> = {
   blocksize: 4,
   mainParameter: 'blocksize',
-  fragmentSource: `
-    precision highp float;
-    uniform sampler2D uTexture;
-    uniform float uBlocksize;
-    uniform float uStepW;
-    uniform float uStepH;
-    varying vec2 vTexCoord;
-    void main() {
-      float blockW = uBlocksize * uStepW;
-      float blockH = uBlocksize * uStepW;
-      int posX = int(vTexCoord.x / blockW);
-      int posY = int(vTexCoord.y / blockH);
-      float fposX = float(posX);
-      float fposY = float(posY);
-      vec2 squareCoords = vec2(fposX * blockW, fposY * blockH);
-      vec4 color = texture2D(uTexture, squareCoords);
-      gl_FragColor = color;
-    }
-    `,
 };
 
 /**
@@ -74,6 +56,10 @@ export class Pixelate extends BaseFilter {
    **/
   isNeutralState() {
     return this.blocksize === 1;
+  }
+
+  protected getFragmentSource(): string {
+    return fragmentSource;
   }
 
   /**
