+++ src/filters/invert.shaders.ts
@@ -1,0 +1,19 @@
+export const fragmentSource = `
+  precision highp float;
+  uniform sampler2D uTexture;
+  uniform int uInvert;
+  uniform int uAlpha;
+  varying vec2 vTexCoord;
+  void main() {
+    vec4 color = texture2D(uTexture, vTexCoord);
+    if (uInvert == 1) {
+      if (uAlpha == 1) {
+        gl_FragColor = vec4(1.0 - color.r,1.0 -color.g,1.0 -color.b,1.0 -color.a);
+      } else {
+        gl_FragColor = vec4(1.0 - color.r,1.0 -color.g,1.0 -color.b,color.a);
+      }
+    } else {
+      gl_FragColor = color;
+    }
+  }
+`;
