+++ src/filters/pixelate.shaders.ts
@@ -1,0 +1,19 @@
+export const fragmentSource = `
+  precision highp float;
+  uniform sampler2D uTexture;
+  uniform float uBlocksize;
+  uniform float uStepW;
+  uniform float uStepH;
+  varying vec2 vTexCoord;
+  void main() {
+    float blockW = uBlocksize * uStepW;
+    float blockH = uBlocksize * uStepW;
+    int posX = int(vTexCoord.x / blockW);
+    int posY = int(vTexCoord.y / blockH);
+    float fposX = float(posX);
+    float fposY = float(posY);
+    vec2 squareCoords = vec2(fposX * blockW, fposY * blockH);
+    vec4 color = texture2D(uTexture, squareCoords);
+    gl_FragColor = color;
+  }
+`;
